Indicate when an interactive figure is being shown
Print 'Showing plot...' to stdout when showing an interactive matplotlib figure.  I simply searched for "show()" in the code to find appropriate places for this insertion.  I've been caught out before when PyWiFeS apparently stops, but in fact it's just showing me a figure that is hidden by another window.